publish_findings/requirements.txt
google-cloud-bigquery==2.6.1
google-cloud-securitycenter==1.1.0